Hotfix: Allow items from Abstract Factories to have setShared() called
Some items that are created from Abstract Factories, we may want to
re-create each time - rather than having a shared object.  This allows
checks whether the requested object is able to be created via an
Abstract Factory.  The rest of the code is already in place to allow
this not to be added to $sm->instances;