AbstractPluginManager should cache by default
Previous implementation proxied to `build()` always, which made it
impossible to have a caching plugin manager. The new implementation
proxies to `parent::get()` if no options are passed, and to `build()`
when options are present (i.e., presence of options always disables
caching). This allows implementors to set the `$sharedByDefault`
property if they *never* want cached instances.