Add method for retrieving configuration to ConfigInterface
Added `toArray()` method to `ConfigInterface` to provide a mechanism for
fetching configuration without needing to configure the service manager. This
will be useful in particular for `Zend\ModuleManager\Listener\ServiceListener`,
which primarily needs to aggregate and merge configuration *prior* to creating a
new instance of a plugin manager.

An implementation was added to the `Config` class, along with tests.