Cache by requested service name
Previous behavior of the service manager allowed specifying aliases when
creating the `shared` configuration. Prior to this patch, that did not
work with the refactored code.

This patch updates the code to check for a cached service under the
_requested_ service name, not the _resolved_ service name, which allows
specifying sharing configuration based on aliases.

Internally, it caches up to twice:

- If the resolved service name is supposed to be shared, it will cache
  under the resolved name.
- If the requested service name is supposed to be shared, it will cache
  under the resolved name.

This allows having separate instances based on service name.