Ensure plugin manager creation context travels with clone
Prior to this patch, calling `withConfig()` on a plugin manager would
reset the `creationContext` to the plugin manager itself — instead of
retaining the original creation context. This patch updates
`ServiceManager::withConfig()` to ensure that if the creation context is
different than the original instance, it gets injected into the clone.