Move `getServiceLocator()` to service manager
Since the container passed to factories is always the creation context
container in v3, `getServiceLocator()` will always be the service
manager itself, not a plugin manager. By pushing the method into the
service manager, we achieve backwards compatibility with v2; by marking
it deprecated, we can remove it in the future.