Adds the InvokableFactory
Most of the component refactors are complete, and will need to be re-done.
However, one item that should likely be retained is usage of the `InvokableFactory`
for invokable instances. This patch adds that factory to the v2 series,
having it implement `Zend\ServiceManager\FactoryInterface`, but retaining the
defined invokable method from v3. It raises an exception if the `$requestedName`
is not provided, ensuring that any tests against the functionality are using
a `ServiceManager` instance.