Fix behavior of InvokableFactory when aliases are used
v2 displays 2 distinct behaviors when invoking a factory:

- For normal factory mappings, `$requestedName` will be the name under
  which the service was requested, and `$canonicalName` will be the
  normalized name.
- If an alias was used, and resolved to a factory, the `$canonicalName`
  will be the resolved name, while `$requestedName` will be the name
  used to retrieve the service.

In the second case, when attempting to provide forwards compatibility
with v3, `$canonicalName` will typically be a fully qualified class
name. If it is (which typically will never be the case with
canonicalization), it should be the name used to create the instance.

Otherwise, `$requestedName` should be used, if the class exists.