Allow creation options as passed via v2 factories.
In v2, options passed via the second argument to `AbstractPluginManager::get()`
are passed to factory *constructors*; the factory must then determine
whether or not to use them when creating the object instance.

This patch updates `InvokableFactory` to add a constructor that can
accept creation arguments; `createService()` will then pass them to
`__invoke()` when present.