Ports zend-mvc LazyControllerAbstractFactory to zend-servicemanager
This patch ports zend-mvc's `LazyControllerAbstractFactory` to a more
generic `ReflectionBasedAbstractFactory` that can be used anywhere. The
principal change is that it does not restrict type generation to only
dispatchables.

One additional change was made: the zend-mvc version hard-codes in a
number of "well known" services that use short names, mapping them to
the actual classes that handle them. This patch keeps them, but allows
you to override them via the constructor.