More and better benchmarking of new abstract factories
This patch provides more benchmarking for the new abstract factories, pulling
them each into separate benchmark classes to ensure they operate in isolation to
other abstract factories, and to allow seeing baseline timings.

Additionally, each has two benchmarking suites: one demonstrating usage as an
abstract factory, another demonstrating usage as a mapped factory.

The basic results are that the `ConfigAbstractFactory` operates at roughly the
same speed as a normal abstract factory that contains logic for creating an
instance. The `ReflectionBasedAbstractFactory` is about 100% slower when at
least one dependency is present (and ever-so-slightly faster when no
dependencies are present).