Reintroduced 'factories' configuration.
With ['services' => [ \stdClass::class => new \stdClass::class() ]
instead of 'factories' new errors with other root causes come up.
I'll investigate and test and try to fix asap.

Anyway, as it is now, the fix this PR provides will remain valid.
Hey, please look at the original code. It's quite obvious that the
classification rule for a shared alias should be the same at lines x and
x + 10 or so.