Improved tests aliases with abstract factories.
There was an error in the testcase `testAbstractFactoryShouldBeCheckedForResolvedAliasesInsteadOfAliasName`
where the returnValueMap of phpunit was used wrong. This caused a false positive. The test was returning an
array which is interpreted as `TRUE`.

I discovered this when I added the `testResolvedAliasFromAbstractFactory`, which should do the opposite form the
earlier mentioned testcase.

Beside that I found a way that the servicemanager was returning `null` instead of returning false when the last fallback
on abstract factories failed.